refactor: move ValueName from message.py to value.py
