+++ tests/test_controller.py
@@ -3,7 +3,8 @@ import numpy as np
 import pytest
 import random
 
-from vent.common.message import SensorValues, ControlSetting, Alarm, AlarmSeverity, ValueName
+from vent.common.message import SensorValues, ControlSetting, Alarm, AlarmSeverity
+from vent.common.values import ValueName
 from vent.coordinator.coordinator import get_coordinator
 from vent.controller.control_module import get_control_module
